requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.0.3
djangorestframework==3.9.0
html5lib==1.0.1
idna==2.7
jsonfield2==3.0.1
Pillow==5.3.0
pyPdf==1.13
PyPDF2==1.26.0
pytz==2018.7
reportlab==3.5.12
requests==2.20.1
six==1.11.0
tika==1.19
urllib3==1.24.1
webencodings==0.5.1
xhtml2pdf==0.2.3
